Move codes to where it should be located
